tools/cex_collector/pyproject.toml
[project]
name = "grok-cli"
version = "0.1.0"
description = "极简的 Grok CLI 工具，支持 web_search 和 x_search"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "xai-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
grok = "main:main"
